Split TestRuleOnCode into a source and header file
